use factory instead (tests still fail)
